Fixed refreshing and retain ticket information on refresh
